Code changes 11/17/25
Changed update batch size to be larger to reduce stream processing delay, both before and after calibration

Added another plot that displays the predicted Y in real time

Need to alter our calibration procedure making the dots a little easier to track and to avoid noise.

If you need to make the changes on windows just copy this file and use the keyboard module for event listening.